+++ src/Cine.Core/iado.cs
@@ -19,5 +19,5 @@ public interface IAdo
 
     void AltaEntrada(Entrada entrada);
     List<Entrada> ObtenerEntrada();
-    public Proyeccion ProyeccionPorId(byte id);
+    public Proyeccion ProyeccionPorId(ushort id);
 }
